requirements.txt
# NautilusTrader DeepSeek AI Strategy Dependencies
# Python >= 3.12 required

# Core framework - 1.224.0 includes reconciliation loop fix, trailing stop LAST_PRICE fix, algo order fixes
nautilus_trader==1.224.0

# AI and API clients
openai>=1.0.0
requests>=2.31.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Performance Metrics (v3.0.1 - use official libraries)
# See: docs/research/EVALUATION_FRAMEWORK.md
empyrical-reloaded>=0.5.12  # Sharpe/Sortino/Calmar/MDD/VaR (Python 3.12 compatible fork)
scipy>=1.9.0                # Bootstrap, statistical tests
statsmodels>=0.14.0         # Multiple hypothesis testing correction

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Cache and persistence
redis>=5.0.0

# Optional: For enhanced logging
rich>=13.0.0

# Notifications and monitoring
python-telegram-bot>=21.0

# HTTP client
httpx>=0.27.0
aiohttp>=3.9.0

# Binance REST API (for E2E tests and direct API access)
binance-futures-connector>=4.0.0

# Testing
pytest>=7.0.0

# =============================================================================
# v2.0 Upgrade Plan Dependencies
# Ref: docs/upgrade_plan_v2/00_INDEX_v3.md
# =============================================================================

# Phase 0 — Baseline measurement
quantstats>=0.0.62              # KPI baseline (Sharpe/Calmar/Monte Carlo)

# Phase 1 — Code hardening
tenacity>=9.0.0                 # REST client unified retry
market-profile>=0.3.0           # Volume Profile (VPOC/VAH/VAL)
pandas-ta>=0.3.14b              # ATR/SMA/BB/Pivot standard implementations
pandera>=0.20.0                 # DataFrame schema validation
hmmlearn>=0.3.2                 # HMM regime detection (4-state)
scikit-learn>=1.5.0             # Required by hmmlearn + DBSCAN S/R clustering

# Phase 2 — Position + risk + parameter optimization
optuna>=4.0.0                   # Walk-forward parameter optimization
prometheus-client>=0.21.0       # Metrics export for Grafana
